+++ apps/docs/src/app/platform/component-docs/platform-wizard-generator/examples/wizard-generator-customizable-example.component.ts
@@ -139,7 +139,7 @@ export class WizardGeneratorCustomizableExampleComponent implements OnDestroy {
                 items: this.stepItems,
                 appendToWizard: false,
                 displaySummaryStep: true,
-                responsivePaddings: false,
+                responsivePaddings: true,
                 title: this.wizardTitle,
                 goNextButtonTemplate: this.goNextTemplate,
                 goBackButtonTemplate: this.goBackTemplate,