+++ libs/platform/src/lib/components/wizard-generator/base-wizard-generator.ts
@@ -46,7 +46,7 @@ export class BaseWizardGenerator implements OnDestroy {
      * @description Whether or not apply responsive paddings styling.
      */
      @Input()
-     responsivePaddings = false;
+     responsivePaddings = true;
 
     /**
      * @description Button labels to be used in Wizard navigation